@@ -1,6 +1,6 @@
 module github.com/ClusterLabs/ha_cluster_exporter
 
-go 1.20
+go 1.23
 
 require (
 	github.com/alecthomas/kingpin/v2 v2.4.0
